investment-backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0
django-cors-headers>=4.3.1

